requirements.txt
babel==2.17.0
beautifulsoup4==4.13.4
certifi==2025.7.14
charset-normalizer==3.4.2
courlan==1.3.2
dateparser==1.2.2
feedparser==6.0.11
htmldate==1.9.3
idna==3.10
jusText==3.0.2
lxml==5.4.0
packaging==25.0
pillow==11.3.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2025.7.33
requests==2.32.4
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.7
tld==0.13.1
trafilatura==2.0.0
typing_extensions==4.14.1
tzlocal==5.3.1
urllib3==2.5.0

#OCR dependencies

pytesseract>=0.3.10
pdf2image>=1.17.0
Pillow>=9.0.0
numpy>=1.21.0
opencv-python>=4.5.0    
# Tesseract and poppler are system dependencies, not Python packages, brew install tesseract-ocr and poppler